# Move compile timing to inside `@time`'s main timing block. Fixes >100% compilation time reports