--- a/base/timing.jl
+++ b/base/timing.jl
@@ -204,15 +204,14 @@ julia> @time begin
 macro time(ex)
     quote
         while false; end # compiler heuristic: compile this block (alter this if the heuristic changes)
+
+        ## ensure these are compiled as they are first called within the compilation-timed region
+        precompile(time_ns, ())
+        precompile(cumulative_compile_time_ns_after, ())
+
         local stats = gc_num()
         local compile_elapsedtime = cumulative_compile_time_ns_before()
         local elapsedtime = time_ns()
-        ## ensure time samplers are compiled
-        compile_elapsedtime = cumulative_compile_time_ns_after() - compile_elapsedtime
-        elapsedtime = time_ns() - elapsedtime
-        ## reset timers
-        compile_elapsedtime = cumulative_compile_time_ns_before()
-        elapsedtime = time_ns()
         local val = $(esc(ex))
         elapsedtime = time_ns() - elapsedtime
         compile_elapsedtime = cumulative_compile_time_ns_after() - compile_elapsedtime
@@ -256,15 +255,14 @@ pool allocs:       1
 macro timev(ex)
     quote
         while false; end # compiler heuristic: compile this block (alter this if the heuristic changes)
+
+        ## ensure these are compiled as they are first called within the compilation-timed region
+        precompile(time_ns, ())
+        precompile(cumulative_compile_time_ns_after, ())
+
         local stats = gc_num()
         local compile_elapsedtime = cumulative_compile_time_ns_before()
         local elapsedtime = time_ns()
-        ## ensure time samplers are compiled
-        compile_elapsedtime = cumulative_compile_time_ns_after() - compile_elapsedtime
-        elapsedtime = time_ns() - elapsedtime
-        ## reset timers
-        compile_elapsedtime = cumulative_compile_time_ns_before()
-        elapsedtime = time_ns()
         local val = $(esc(ex))
         elapsedtime = time_ns() - elapsedtime
         compile_elapsedtime = cumulative_compile_time_ns_after() - compile_elapsedtime