--- a/base/timing.jl
+++ b/base/timing.jl
@@ -204,11 +204,6 @@ julia> @time begin
 macro time(ex)
     quote
         while false; end # compiler heuristic: compile this block (alter this if the heuristic changes)
-
-        ## ensure these are compiled as they are first called within the compilation-timed region
-        precompile(time_ns, ())
-        precompile(cumulative_compile_time_ns_after, ())
-
         local stats = gc_num()
         local elapsedtime = time_ns()
         local compile_elapsedtime = cumulative_compile_time_ns_before()
@@ -255,11 +250,6 @@ pool allocs:       1
 macro timev(ex)
     quote
         while false; end # compiler heuristic: compile this block (alter this if the heuristic changes)
-
-        ## ensure these are compiled as they are first called within the compilation-timed region
-        precompile(time_ns, ())
-        precompile(cumulative_compile_time_ns_after, ())
-
         local stats = gc_num()
         local elapsedtime = time_ns()
         local compile_elapsedtime = cumulative_compile_time_ns_before()