@IanButterworth 2021-07-06
@chriselrod that seems like a different issue, perhaps Distributed-related. Could different processes be using the same thread-specific compile timers? I didn't think so, but I don't have another explanation.

This PR is more about the slightly > 100% reports